@@ -9,18 +9,18 @@ name = "pyhmy"
 version = "0.1.2"
 description = "A library for interacting and working the Harmony blockchain and related codebases"
 readme = "README.md"
-license = { text = "MIT" }
+license = "MIT"
 keywords = [ "Harmony", "blockchain", "protocol", "staking" ]
 dependencies = [
-    "pexpect",
-    "requests",
-    "eth-rlp",
-    "eth-account >= 0.5.5",
-    "eth-utils",
-    "hexbytes",
-    "cytoolz"
+    "pexpect == 4.9.0",
+    "requests == 2.32.3",
+    "eth-rlp == 2.2.0",
+    "eth-account == 0.13.7",
+    "eth-utils == 5.3.0",
+    "hexbytes == 1.3.0",
+    "cytoolz == 1.0.1"
 ]
-requires-python = ">=3.0"
+requires-python = ">=3.11.2"
 
 [project.optional-dependencies]
 dev = [ "black", "autopep8", "yapf", "twine", "build", "docformatter", "bumpver" ]
