@@ -5,7 +5,7 @@ import { Transaction, RLPSign, TxStatus } from '@harmony-js/transaction';
 import { HttpProvider, Messenger } from '@harmony-js/network';
 import { ChainType, ChainID } from '@harmony-js/utils';
 
-const provider = new HttpProvider('http://localhost:9500');
+const provider = new HttpProvider('http://localhost:9620');
 let privateKey = '4edef2c24995d15b0e25cbd152fb0e2c05d3b79b9c2afd134e6f59f91bf99e48'
 let hmyMessenger = new Messenger(provider, ChainType.Ethereum, ChainID.Default);
 
@@ -42,7 +42,7 @@ def test_eth_transaction():
         "4edef2c24995d15b0e25cbd152fb0e2c05d3b79b9c2afd134e6f59f91bf99e48",
     )
     assert (
-        signed_tx.rawTransaction.hex() ==
+        signed_tx.raw_transaction.to_0x_hex() ==
         "0xf85d0201649414791697260e4c9a71f18484c9f997b308e5932505801ca0b364f4296bfd3231889d1b9ac94c68abbcb8ee6a6c7a5fa412ac82b5b7b0d5d1a02233864842ab28ee4f99c207940a867b0f8534ca362836190792816b48dde3b1"
     )
 
@@ -53,7 +53,7 @@ import { Transaction, RLPSign, TxStatus } from '@harmony-js/transaction';
 import { HttpProvider, Messenger } from '@harmony-js/network';
 import { ChainType, ChainID } from '@harmony-js/utils';
 
-const provider = new HttpProvider('http://localhost:9500');
+const provider = new HttpProvider('http://localhost:9620');
 let privateKey = '4edef2c24995d15b0e25cbd152fb0e2c05d3b79b9c2afd134e6f59f91bf99e48'
 let hmyMessenger = new Messenger(provider, ChainType.Harmony, ChainID.HmyMainnet);
 
@@ -95,7 +95,7 @@ def test_hmy_transaction():
         "4edef2c24995d15b0e25cbd152fb0e2c05d3b79b9c2afd134e6f59f91bf99e48",
     )
     assert (
-        signed_tx.rawTransaction.hex() ==
+        signed_tx.raw_transaction.to_0x_hex() ==
         "0xf85f02016480019414791697260e4c9a71f18484c9f997b308e59325058026a02a203357ca6d7cdec981ad3d3692ad2c9e24536a9b6e7b486ce2f94f28c7563ea010d38cd0312a153af0aa7d8cd986040c36118bba373cb94e3e86fd4aedce904d"
     )
 
@@ -127,6 +127,6 @@ def test_hmy_eth_compatible_transaction():
         "0x1111111111111111111111111111111111111111111111111111111111111111",
     )
     assert (
-        signed_tx.rawTransaction.hex() ==
+        signed_tx.raw_transaction.to_0x_hex() ==
         "0xf8748085174876e80082520880019419e7e376e7c213b7e7e7e46cc70a5dd086daff2a880de0b6b3a764000082abcd84c6ac98a4a0ef05b0bc0827ddd046ee8196253c6609484e28b8f250d860736ce8e053469f04a07229a88ce68241a9e77dcb1592a67a9a6a215d855a784467c9f550f8c4cb9b76"
     )
