@@ -24,7 +24,7 @@ def test_instantiate_validator( setup_blockchain ):
 
 def test_load_validator( setup_blockchain ):
     if not test_validator_object:
-        pytest.skip( "Validator not instantiated yet" )
+        pytest.fail( "Validator not instantiated yet" )
     info = {
         "name": "Alice",
         "identity": "alice",
@@ -80,14 +80,14 @@ console.log(signed)
 
 def test_create_validator_sign( setup_blockchain ):
     if not ( test_validator_object or test_validator_loaded ):
-        pytest.skip( "Validator not instantiated yet" )
+        pytest.fail( "Validator not instantiated yet" )
     signed_hash = test_validator_object.sign_create_validator_transaction(
         2,
         int( convert_one_to_atto( 1 ) ),
         100,
         "4edef2c24995d15b0e25cbd152fb0e2c05d3b79b9c2afd134e6f59f91bf99e48",
         2,
-    ).rawTransaction.hex()
+    ).raw_transaction.to_0x_hex()
     assert (
         signed_hash ==
         "0xf9017580f9012394ebcd16e8c1d8f493ba04e99a56474122d81a9c58f83885416c69636585616c69636591616c6963652e6861726d6f6e792e6f6e6583426f6295446f6e2774206d6573732077697468206d65212121dcc8872386f26fc10000c9880c7d713b49da0000c887b1a2bc2ec500008a021e19e0c9bab24000008a0878678326eac9000000f1b030b2c38b1316da91e068ac3bd8751c0901ef6c02a1d58bc712104918302c6ed03d5894671d0c816dad2b4d303320f202f862b86068f800b6adf657b674903e04708060912b893b7c7b500788808247550ab3e186e56a44ebf3ca488f8ed1a42f6cef3a04bd5d2b2b7eb5a767848d3135b362e668ce6bba42c7b9d5666d8e3a83be707b5708e722c58939fe9b07c170f3b70624148a021e27c1806e59a4000002880de0b6b3a76400006428a0c6c7e62f02331df0afd4699ec514a2fc4548c920d77ad74d98caeec8c924c09aa02b27b999a724b1d341d6bbb0e877611d0047542cb7e380f9a6a272d204b450cd"
@@ -139,7 +139,7 @@ console.log(signed)
 
 def test_edit_validator_sign( setup_blockchain ):
     if not ( test_validator_object or test_validator_loaded ):
-        pytest.skip( "Validator not instantiated yet" )
+        pytest.fail( "Validator not instantiated yet" )
     signed_hash = test_validator_object.sign_edit_validator_transaction(
         2,
         int(convert_one_to_atto(1)),
@@ -150,7 +150,7 @@ def test_edit_validator_sign( setup_blockchain ):
         "0x68f800b6adf657b674903e04708060912b893b7c7b500788808247550ab3e186e56a44ebf3ca488f8ed1a42f6cef3a04bd5d2b2b7eb5a767848d3135b362e668ce6bba42c7b9d5666d8e3a83be707b5708e722c58939fe9b07c170f3b7062414",  # add key sig
         "4edef2c24995d15b0e25cbd152fb0e2c05d3b79b9c2afd134e6f59f91bf99e48",
         2,
-    ).rawTransaction.hex()
+    ).raw_transaction.to_0x_hex()
     assert (
         signed_hash ==
         "0xf9018401f9013294ebcd16e8c1d8f493ba04e99a56474122d81a9c58f83885416c69636585616c69636591616c6963652e6861726d6f6e792e6f6e6583426f6295446f6e2774206d6573732077697468206d65212121c887d529ae9e8600008a021e19e0c9bab24000008a0878678326eac9000000b0b9486167ab9087ab818dc4ce026edb5bf216863364c32e42df2af03c5ced1ad181e7d12f0e6dd5307a73b62247608612b0b9486167ab9087ab818dc4ce026edb5bf216863364c32e42df2af03c5ced1ad181e7d12f0e6dd5307a73b62247608611b86068f800b6adf657b674903e04708060912b893b7c7b500788808247550ab3e186e56a44ebf3ca488f8ed1a42f6cef3a04bd5d2b2b7eb5a767848d3135b362e668ce6bba42c7b9d5666d8e3a83be707b5708e722c58939fe9b07c170f3b706241402880de0b6b3a76400006427a0ecdae4a29d051f4f83dd54004858fbf0f7820e169b8e1846245835ceb686ee12a04b2336eb5830e30720137b2de539518fd5655467fef140ab31fde881a19f256a"
@@ -159,7 +159,7 @@ def test_edit_validator_sign( setup_blockchain ):
 
 def test_invalid_validator( setup_blockchain ):
     if not ( test_validator_object or test_validator_loaded ):
-        pytest.skip( "Validator not instantiated yet" )
+        pytest.fail( "Validator not instantiated yet" )
     with pytest.raises( InvalidValidatorError ):
         info = {
             "name": "Alice",
@@ -191,7 +191,7 @@ def test_invalid_validator( setup_blockchain ):
 
 def test_validator_getters( setup_blockchain ):
     if not ( test_validator_object or test_validator_loaded ):
-        pytest.skip( "Validator not instantiated yet" )
+        pytest.fail( "Validator not instantiated yet" )
     assert (
         test_validator_object.get_address() ==
         "one1a0x3d6xpmr6f8wsyaxd9v36pytvp48zckswvv9"
